requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.7.0
gymnasium==1.1.1
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
litellm==1.75.5.post1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mini-swe-agent==1.8.1
multidict==6.6.4
mypy_extensions==1.1.0
nest_asyncio==1.6.0
numpy==2.0.2
openai==1.99.9
packaging==25.0
pandas==2.3.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==22.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
tenacity==9.1.2
textual==5.3.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.23.2
typer==0.16.0
typing_extensions==4.14.0
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.1
zipp==3.21.0
